+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `CreateDataColumns` to support more values in the `DataTable`.
 - Add `ConsoleTaskExtension` to wait task to finish and show progress bar.
 - Update `AppBundleUtils` with `FindAppBundle` to find `AppBundle` by app name or bundle folder.
+- Update `CACHE_TOTAL_MINUTES` in `DownloadUtils` to 1 second.
 
 ## [1.0.1] / 2025-05-31
 ### Features